mise a jour des commentaires de match

+++ app/Http/Controllers/CommentaireController.php
@@ -94,7 +94,6 @@ class CommentaireController extends Controller
      */
     public function update(Request $request, Commentaire $commentaire)
     {
-        
     }
 
     /**
@@ -105,20 +104,20 @@ class CommentaireController extends Controller
      */
     public function destroy($id)
     {
-        
+
         $commentaire = Commentaire::find($id);
 
-        if($commentaire->statistic->action == "goal" && $commentaire->team_action == "home"){
-            $commentaire->delete();
-            $commentaire->commentator->match->home_score -= 1;
-            $commentaire->commentator->match->save();
-        }
-        elseif($commentaire->statistic->action == "goal" && $commentaire->team_action == "away"){
-            $commentaire->delete();
-            $commentaire->commentator->match->away_score -= 1;
-            $commentaire->commentator->match->save();
-        }
-        else{
+        if ($commentaire->statistic) {
+            if ($commentaire->statistic->action == "goal" && $commentaire->team_action == "home") {
+                $commentaire->delete();
+                $commentaire->commentator->match->home_score -= 1;
+                $commentaire->commentator->match->save();
+            } elseif ($commentaire->statistic->action == "goal" && $commentaire->team_action == "away") {
+                $commentaire->delete();
+                $commentaire->commentator->match->away_score -= 1;
+                $commentaire->commentator->match->save();
+            }
+        } else {
             $commentaire->delete();
         }
